added encyption feature
