+++ appinfo/info.xml
@@ -21,8 +21,11 @@
 	</dependencies>
     <settings>
 		<personal>OCA\NMCSettings\Settings\Personal\NmcPersonalInfo</personal>
-		<personal-section>OCA\NMCSettings\Sections\Personal\AccountSettings</personal-section>
+		<personal>OCA\NMCSettings\Settings\Personal\NmcPersonalInfoStorage</personal>
 		<personal>OCA\NMCSettings\Settings\Personal\Security\NmcAuthtokens</personal>
+		<personal>OCA\NMCSettings\Settings\Personal\Security\NmcDevices</personal>
+		<personal>OCA\NMCSettings\Settings\Personal\Security\NmcEncryption</personal>
+		<personal-section>OCA\NMCSettings\Sections\Personal\AccountSettings</personal-section>
 		<personal-section>OCA\NMCSettings\Sections\Personal\DevicesSessions</personal-section>
 		<personal>OCA\NMCSettings\Settings\Groupware</personal>
 		<personal>OCA\NMCSettings\Settings\PersonalInfo</personal>
