+++ lib/Settings/Personal/NmcPersonalInfo.php
@@ -98,41 +98,12 @@ class NmcPersonalInfo implements ISettings {
 		$user = $this->userManager->get($uid);
 		$account = $this->accountManager->getAccount($user);
 
-		$imageMimetypes = "'image','image/jpg','image/jpeg','image/gif','image/png','image/svg+xml','image/webp'";
-		$videoMimetypes = "'video','video/3gpp','video/mp4', 'video/mov', 'video/avi', 'video/flv'";
-
-		$imageStorageInBytes = $this->storageUtilization($uid, $imageMimetypes);
-		$videoStorageInBytes = $this->storageUtilization($uid, $videoMimetypes);
-		$photoVideoSizeInBytes = $imageStorageInBytes + $videoStorageInBytes;
-
-		// make sure FS is setup before querying storage related stuff...
-		\OC_Util::setupFS($user->getUID());
-
-		$storageInfo = \OC_Helper::getStorageInfo('/');
-		if ($storageInfo['quota'] === FileInfo::SPACE_UNLIMITED) {
-			$totalSpace = $this->l->t('Unlimited');
-		} else {
-			$totalSpace = \OC_Helper::humanFileSize($storageInfo['total']);
-		}
-
 		$messageParameters = $this->getMessageParameters($account);
-
-		$trashSizeinBytes = self::getTrashbinSize($uid);
-		$filesSizeInBytes = $storageInfo['used'] - ($photoVideoSizeInBytes);
-
-		if($filesSizeInBytes < 0) {
-			$filesSizeInBytes = 0;
-		}
 
 		$personalInfoParameters = [
 			'userId' => $uid,
 			'avatar' => $this->getProperty($account, IAccountManager::PROPERTY_AVATAR),
 			'groups' => $this->getGroups($user),
-			'quota' => $storageInfo['quota'],
-			'totalSpace' => $totalSpace,
-			'tariff' => $this->getTariff($storageInfo['quota']),
-			'usage' => \OC_Helper::humanFileSize($storageInfo['used']),
-			'usageRelative' => round($storageInfo['relative']),
 			'displayName' => $this->getProperty($account, IAccountManager::PROPERTY_DISPLAYNAME),
 			'emailMap' => $this->getEmailMap($account),
 			'phone' => $this->getProperty($account, IAccountManager::PROPERTY_PHONE),
@@ -148,12 +119,6 @@ class NmcPersonalInfo implements ISettings {
 			'role' => $this->getProperty($account, IAccountManager::PROPERTY_ROLE),
 			'headline' => $this->getProperty($account, IAccountManager::PROPERTY_HEADLINE),
 			'biography' => $this->getProperty($account, IAccountManager::PROPERTY_BIOGRAPHY),
-			'trashSize' => \OC_Helper::humanFileSize($trashSizeinBytes),
-			'photoVideoSize' => \OC_Helper::humanFileSize($photoVideoSizeInBytes),
-			'filesSize' => \OC_Helper::humanFileSize($filesSizeInBytes),
-			'trashSizeInPer' => round(($trashSizeinBytes / $storageInfo['quota']) * 100) ,
-			'photoVideoSizeInPer' => round(($photoVideoSizeInBytes / $storageInfo['quota']) * 100),
-			'filesSizeInPer' => round(($filesSizeInBytes / $storageInfo['quota']) * 100) ,
 		];
 
 		$parameters = [
@@ -186,7 +151,7 @@ class NmcPersonalInfo implements ISettings {
 	}
 
 	public function getPriority(): int {
-		return 10;
+		return 0;
 	}
 
 	/**
@@ -338,73 +303,5 @@ class NmcPersonalInfo implements ISettings {
 			$messageParameters[$property . 'Message'] = $message;
 		}
 		return $messageParameters;
-	}
-
-	private function getTariff($quota) {
-
-		$totalSpaceInGB = null;
-
-		if($quota >= 1024) {
-			$totalSpaceInKB = round($quota / 1024, 1);
-			$totalSpaceInMB = round($totalSpaceInKB / 1024, 1);
-			$totalSpaceInGB = round($totalSpaceInMB / 1024, 1);
-		}
-
-		if ($quota == 0) {
-			$tariff = $this->l->t('No space allocated');
-		} elseif($quota === FileInfo::SPACE_UNLIMITED) {
-			$tariff = $this->l->t('Unlimited');
-		} elseif($quota === FileInfo::SPACE_UNKNOWN) {
-			$tariff = $this->l->t('Space unknown');
-		} elseif($quota === FileInfo::SPACE_NOT_COMPUTED) {
-			$tariff = $this->l->t('Space not computed');
-		} elseif ($totalSpaceInGB == 1 || $totalSpaceInGB == 3 || $totalSpaceInGB == 10) {
-			$tariff = $this->l->t('MagentaCLOUD Free');
-		} elseif ($totalSpaceInGB == 15 || $totalSpaceInGB == 25) {
-			$tariff = $this->l->t('MagentaCLOUD S');
-		} elseif ($totalSpaceInGB == 100) {
-			$tariff = $this->l->t('MagentaCLOUD M');
-		} elseif ($totalSpaceInGB == 500) {
-			$tariff = $this->l->t('MagentaCLOUD L');
-		} elseif ($totalSpaceInGB == 1024) {
-			$tariff = $this->l->t('MagentaCLOUD XL');
-		} elseif ($totalSpaceInGB == 5120) {
-			$tariff = $this->l->t('MagentaCLOUD XXL');
-		} else {
-			$tariff = $this->l->t('Tariff unknown');
-		}
-
-		return $tariff;
-	}
-
-	private static function getTrashbinSize($user) {
-		$view = new View('/' . $user);
-		$fileInfo = $view->getFileInfo('/files_trashbin');
-		return isset($fileInfo['size']) ? $fileInfo['size'] : 0;
-	}
-
-	private function storageUtilization($user = null, $filterMimetypes = null) {
-		$details = null;
-
-		$rootFolder = \OC::$server->getRootFolder()->getUserFolder($user);
-		$storageId = $rootFolder->getStorage()->getCache()->getNumericStorageId();
-
-		$query = $this->db->getQueryBuilder();
-
-		$query->selectAlias($query->func()->sum('size'), 'f1')
-			->from('filecache', 'fc')
-			->innerJoin('fc', 'mimetypes', 'mt', $query->expr()->eq('fc.mimetype', 'mt.id'))
-			->where('mt.mimetype in('.$filterMimetypes.')')
-			->andWhere($query->expr()->neq('fc.size', $query->createPositionalParameter(-1)))
-			->andWhere("fc.path NOT Like 'files_trashbin/files/%'")
-			->andWhere($query->expr()->eq('fc.storage', $query->createPositionalParameter($storageId)));
-
-		$result = $query->execute();
-
-		while ($row = $result->fetch()) {
-			$details = $row['f1'];
-		}
-		$result->closeCursor();
-		return $details;
 	}
 }
