+++ webpack.config.cjs
@@ -6,6 +6,7 @@ const webpackConfig = require('@nextcloud/webpack-vue-config')
 webpackConfig.entry = {
 	    ...webpackConfig.entry, 
 		nmcsettings: path.join(__dirname, 'src', 'js', 'nmcsettings.js'),
+		devices: path.join(__dirname, 'src', 'js', 'devices.js'),
 		personal: path.join(__dirname, 'src', 'js', 'personal.js'),
 		sessions: path.join(__dirname, 'src', 'js', 'sessions.js'),
 	}
